feat: add @types/mdx dependency and refactor blog utility functions
