+++ package.json
@@ -56,6 +56,7 @@
     "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/typography": "^0.5.10",
     "@types/marked": "^5.0.2",
+    "@types/mdx": "^2.0.13",
     "@types/mocha": "^10.0.4",
     "@types/node": "^20.9.0",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
